+++ test_app/lib/test_app_web/router.ex
@@ -1,5 +1,6 @@
 defmodule TestAppWeb.Router do
   use TestAppWeb, :router
+  import Phoenix.LiveView.Router
 
   pipeline :api do
     plug :accepts, ["json"]
@@ -10,5 +11,11 @@ defmodule TestAppWeb.Router do
 
     get "/health", HealthController, :show
     post "/counter/increment", CounterController, :increment
+  end
+
+  scope "/", TestAppWeb do
+    pipe_through :api
+
+    live "/lv", TestLive, :show
   end
 end